further compartmentalized front facing headers
 - all front facing headers with needed internal resources include them
