+++ engine/headers/internal/audio.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "../audio.h"
 
 /**
