+++ engine/headers/internal/camera.h
@@ -1,3 +1,4 @@
+#pragma once
 /**
  * INTENRAL USE ONLY
  * methods for engine internal uses of camera updates
