+++ engine/headers/internal/entitymanager.h
@@ -1,7 +1,4 @@
 #pragma once
-#include "entity.h"
-#include <vector>
-
 
 namespace entities {
     void update();
