+++ engine/headers/internal/handler.h
@@ -1,3 +1,4 @@
+#pragma once
 namespace hnd {
     void render();
     void update();
