+++ engine/headers/sprite.h
@@ -1,30 +1,7 @@
 #pragma once
 #include "alignment.h"
 #include <string>
-#include <memory>
-
-struct Texture {
-private:
-	std::string path;
-    int loaded;
-
-    void lazyload();
-    void unload();
-    
-public:
-	int w, h, frames, delta;
-	SDL_Texture* sheet;
-	SDL_Rect** clips;
-
-	Texture(std::string path) : path(path) {
-        loaded = 0;
-    }
-
-    Texture() : path(NULL) {}
-
-    void update();
-    void ping();
-};
+#include <internal/renderable.h>
 
 struct Renderable {
 public:
